requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile
#
absl-py==1.2.0
    # via
    #   tensorboard
    #   tensorflow
apscheduler==3.9.1
    # via -r requirements.in
astunparse==1.6.3
    # via tensorflow
attrs==22.1.0
    # via
    #   jsonschema
    #   pytest
beautifulsoup4==4.11.1
    # via webtest
cachetools==5.2.0
    # via google-auth
certifi==2022.9.24
    # via requests
charset-normalizer==2.1.1
    # via requests
click==8.1.3
    # via
    #   clickclick
    #   flask
clickclick==20.10.2
    # via connexion
commonmark==0.9.1
    # via rich
connexion==2.14.1
    # via -r requirements.in
convertdate==2.4.0
    # via holidays
dataclasses==0.6
    # via -r requirements.in
deprecated==1.2.13
    # via limits
flask==2.2.2
    # via
    #   -r requirements.in
    #   connexion
    #   flask-admin
    #   flask-cors
    #   flask-jwt-extended
    #   flask-limiter
flask-admin==1.6.0
    # via -r requirements.in
flask-cors==3.0.10
    # via -r requirements.in
flask-jwt-extended==4.4.4
    # via -r requirements.in
flask-limiter==2.6.3
    # via -r requirements.in
flatbuffers==2.0.7
    # via tensorflow
gast==0.4.0
    # via tensorflow
google-auth==2.11.1
    # via
    #   google-auth-oauthlib
    #   tensorboard
google-auth-oauthlib==0.4.6
    # via tensorboard
google-pasta==0.2.0
    # via tensorflow
grpcio==1.49.1
    # via
    #   tensorboard
    #   tensorflow
gunicorn==20.1.0
    # via -r requirements.in
h5py==3.7.0
    # via
    #   pvlib
    #   tensorflow
hijri-converter==2.2.4
    # via holidays
holidays==0.16
    # via -r requirements.in
idna==3.4
    # via requests
inflection==0.5.1
    # via connexion
influxdb==5.3.1
    # via -r requirements.in
iniconfig==1.1.1
    # via pytest
isodate==0.6.1
    # via rdflib
itsdangerous==2.1.2
    # via
    #   connexion
    #   flask
jinja2==3.1.2
    # via
    #   flask
    #   swagger-ui-bundle
joblib==1.2.0
    # via scikit-learn
jsonschema==4.16.0
    # via
    #   -r requirements.in
    #   connexion
keras==2.10.0
    # via tensorflow
keras-preprocessing==1.1.2
    # via tensorflow
korean-lunar-calendar==0.3.1
    # via holidays
libclang==14.0.6
    # via tensorflow
limits==2.7.0
    # via flask-limiter
markdown==3.4.1
    # via tensorboard
markupsafe==2.1.1
    # via
    #   jinja2
    #   werkzeug
    #   wtforms
mongomock==4.1.2
    # via -r requirements.in
msgpack==1.0.4
    # via influxdb
numpy==1.23.3
    # via
    #   -r requirements.in
    #   h5py
    #   keras-preprocessing
    #   opt-einsum
    #   pandas
    #   pvlib
    #   scikit-learn
    #   scipy
    #   tensorboard
    #   tensorflow
oauthlib==3.2.1
    # via requests-oauthlib
opt-einsum==3.3.0
    # via tensorflow
packaging==21.3
    # via
    #   connexion
    #   limits
    #   mongomock
    #   pytest
    #   tensorflow
paho-mqtt==1.6.1
    # via -r requirements.in
pandas==1.5.0
    # via
    #   -r requirements.in
    #   pvlib
pluggy==1.0.0
    # via pytest
protobuf==3.19.5
    # via
    #   tensorboard
    #   tensorflow
pvlib==0.9.3
    # via -r requirements.in
py==1.11.0
    # via pytest
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.2.8
    # via google-auth
pygments==2.13.0
    # via rich
pyjwt==2.5.0
    # via
    #   -r requirements.in
    #   flask-jwt-extended
pymeeus==0.5.11
    # via convertdate
pymongo==4.2.0
    # via -r requirements.in
pyparsing==3.0.9
    # via
    #   packaging
    #   rdflib
pyrsistent==0.18.1
    # via jsonschema
pytest==7.1.3
    # via -r requirements.in
python-dateutil==2.8.2
    # via
    #   -r requirements.in
    #   holidays
    #   influxdb
    #   pandas
pytz==2022.2.1
    # via
    #   apscheduler
    #   influxdb
    #   pandas
    #   pvlib
pytz-deprecation-shim==0.1.0.post0
    # via tzlocal
pyyaml==6.0
    # via
    #   clickclick
    #   connexion
rdflib==6.2.0
    # via -r requirements.in
requests==2.28.1
    # via
    #   connexion
    #   influxdb
    #   pvlib
    #   requests-oauthlib
    #   tensorboard
requests-oauthlib==1.3.1
    # via google-auth-oauthlib
rich==12.5.1
    # via flask-limiter
rsa==4.9
    # via google-auth
scikit-learn==1.1.2
    # via -r requirements.in
scipy==1.9.1
    # via
    #   pvlib
    #   scikit-learn
sentinels==1.0.0
    # via mongomock
six==1.16.0
    # via
    #   apscheduler
    #   astunparse
    #   flask-cors
    #   google-auth
    #   google-pasta
    #   grpcio
    #   influxdb
    #   isodate
    #   keras-preprocessing
    #   python-dateutil
    #   tensorflow
soupsieve==2.3.2.post1
    # via beautifulsoup4
swagger-ui-bundle==0.0.9
    # via -r requirements.in
tensorboard==2.10.1
    # via tensorflow
tensorboard-data-server==0.6.1
    # via tensorboard
tensorboard-plugin-wit==1.8.1
    # via tensorboard
tensorflow==2.10.0
    # via -r requirements.in
tensorflow-estimator==2.10.0
    # via tensorflow
tensorflow-io-gcs-filesystem==0.27.0
    # via tensorflow
termcolor==2.0.1
    # via
    #   -r requirements.in
    #   tensorflow
threadpoolctl==3.1.0
    # via scikit-learn
tomli==2.0.1
    # via pytest
typing-extensions==4.3.0
    # via
    #   flask-limiter
    #   limits
    #   tensorflow
tzdata==2022.4
    # via pytz-deprecation-shim
tzlocal==4.2
    # via apscheduler
urllib3==1.26.12
    # via requests
waitress==2.1.2
    # via webtest
webob==1.8.7
    # via webtest
webtest==3.0.0
    # via -r requirements.in
werkzeug==2.2.2
    # via
    #   -r requirements.in
    #   connexion
    #   flask
    #   flask-jwt-extended
    #   tensorboard
wheel==0.37.1
    # via
    #   astunparse
    #   tensorboard
wrapt==1.14.1
    # via
    #   deprecated
    #   tensorflow
wtforms==3.0.1
    # via flask-admin

# The following packages are considered to be unsafe in a requirements file:
# setuptools
